pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llmeval"
version = "0.1.0"
description = "A pluggable LLM evaluation & benchmarking framework"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
  {name = "Your Name", email = "you@example.com"}
]
keywords = ["llm", "evaluation", "benchmarking", "ai", "nlp"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
  "Programming Language :: Python :: 3.11",
]
dependencies = [
  "anthropic>=0.25.0",
  "openai>=1.30.0",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.2.0",
  "pyyaml>=6.0.1",
  "sentence-transformers>=2.7.0",
  "pandas>=2.2.0",
  "numpy>=1.26.0",
  "rouge-score>=0.1.2",
  "nltk>=3.8.1",
  "aiohttp>=3.9.0",
  "tenacity>=8.3.0",
  "rich>=13.7.0",
  "typer>=0.12.0",
  "python-dotenv>=1.0.0",
  "datasets>=2.19.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.1.0",
  "pytest-asyncio>=0.23.0",
  "pytest-cov>=5.0.0",
  "ruff>=0.4.0",
  "mypy>=1.10.0",
]
dashboard = [
  "streamlit>=1.35.0",
  "plotly>=5.22.0",
]

[project.scripts]
llmeval = "llmeval.cli:app"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
